requirements.txt
# Receipt Importer Requirements

# PDF Processing (choose one or both)
PyPDF2>=3.0.0
PyMuPDF>=1.23.0  # Alternative PDF library (faster, better extraction)
pdfplumber>=0.11.0  # PDF text extraction for Costco receipts (preferred - better layout preservation)
pdfminer.six>=20220524  # PDF text extraction fallback (if pdfplumber unavailable)

# Image processing (if needed for future enhancements)
Pillow>=10.0.0       # Image processing library

# YAML parsing for rule files
PyYAML>=6.0

# Standard libraries (usually included with Python)
# json, logging, pathlib, datetime, re, typing - all built-in

# Web scraping for vendor profiles (Costco, Restaurant Depot online database lookup)
requests>=2.25.1      # HTTP requests to vendor websites
beautifulsoup4>=4.9.3  # HTML parsing for product information

# Excel file processing (for Costco receipts converted to Excel)
pandas>=1.3.0        # Data manipulation and Excel reading
openpyxl>=3.0.0      # Excel file support for pandas

# Database connectivity (for Step 2 and Step 3 - Odoo database operations)
psycopg2-binary==2.9.9  # PostgreSQL adapter for Python (Odoo uses PostgreSQL)
# Note: Version 2.9.9 works with Python 3.8 on macOS. If 2.9.10 causes import errors, use 2.9.9

# AI/LLM Support (optional - for local LLM fallback parsing)
# Note: Supports Ollama (local) and Transformers (local Hugging Face models)
# OpenAI and Gemini APIs have been removed - use local LLMs instead

